refactor: split CLI into main.py, command.py, tui.py (stub)
